Start of static IIIF Image server implementation
This implements the core of the webapp, and the ability to serve
info.json responses.

In principle we don't need an app server to implement the IIIF Image
API from statically-generated tiles, however in practice doing so
affords some advantages.

The main issue I encountered was the need to shard tile folders over
multiple directories, to avoid having directories with unmanageably
large numbers of subdirectories (we currently have half a million
images). Doing this deterministically by sharding image directories
by the prefix of a hash of their identifier is a simple solution,
but achieving this using Apache seems to be impractical (without
exposing the shard prefix in the URL).

The approach I'm going for is to use a very simple app server which
maps IIIF Image request URLs to file paths, responding with a header
containing the location to serve. The front server then takes
responsibility for serving the static file using its optimised static
file serving functionality.

NGINX and Lighthttpd natively support this, and Apache does too with a
3rd party module. This approach is commonly used to offload serving
large file downloads from an app server.

Using Falcon as it appears to be quite performant according to:
    https://github.com/the-benchmarker/web-frameworks
We don't need to do any IO to map requests to locations, so it should
be possible to achieve good performance without much trouble.