Fail if dzi-tiles uses unavailable settings
Some JPEG encoding options require mozjpeg. VIPS itself just warns if
mozjpeg is not available and they're used, but we should fail so that
people can't presume they're using mozjpeg when they're actually not.
This is pretty important given that mozjpeg won't be available by
default, and it's easy to make a mistake when making it available.