Implement input image colour handling
Rather than using assume-srgb as the "umanaged" escape hatch,
"assume-srgb" now just loads the image with the standard sRGB profile,
allowing the result to be mapped to another profile for output. (As a
result we could just get rid of it, and use external-profile with the
sRGB ICC profile...). There's now an explicit "unmanaged" source, which
results in the pixel data being passed through without any
interpretation.