chore: create Docker image for predictable runtime environment
The Dockerfile provides a Python 3.9 environment with libvips built
with MozJPEG support.

MozJPEG is not commonly available from OS package managers (at least
neither Alpine or Debian/Ubuntu provide it). And as a result, the
version of libvips available from package managers won't support
Mozjpeg.

The base target of the Dockerfile is currently used to provide the dev
target, which is intended to be used as a development runtime
environment for the project. It should also be suitable for defining a
runtime image to execute the project in.