test: document libvips behaviour around interpretation of 16 bit images
The libvips changelog for 8.9.0 states that RGB and SRGB
interpretations are synonymous[1], but it's not clear if that applies to
the RGB16 interpretation as well.

These tests show the behaviour in practice. The difference between
RGB16 and SRGB interpretation seems to be in the expected range of the
channel values. libvips does not automatically treat a ushort/16 bit
channel as having 0-0xffff range, but using the RGB16 interpretation
results in that behaviour.

1: https://github.com/libvips/libvips/blob/e5405c8/ChangeLog#L199