fix(docker): enable arithmetic coding in mozpeg build
The issue was that arithmetic cocding is disabled by default for v6b
libjpeg ABI, which is also the default build mode. It's enabled by
default for v8, and we can use that for libvips, so we now build with
v8.

Additionally, Clang is used to compile mozjpeg and libvips. The mozjpeg
build instructions ambiguously states:

> GCC v4.1 (or later) or Clang recommended for best performance

Either possible interpretation recommends Clang for best performance,
so it seems like a sensible choice.