test: xfail cffi callback exception swallow test
It seems that cffi's behaviour has changed. It shouldn't affect the
current implementation, as we intentionally don't raise in the callback,
just the test itself. If this is the case, we could refactor the
implementation, but it requires some investigation to confirm that this
behaviour has changed in all cases.