test: rework xfail cffi callback exception test
A previous commit "test: xfail cffi callback exception swallow test"
changed this test to xfail. It is actually still valid, but upgrading
Python to 3.9 broke it, because in 3.8 Python added the sys.unraisable
hook, and Pytest has a plugin that warns when unraisable exceptions
occur. This caused the expected side-effect of a message on stderr to
not occur. Instead, we now watch for the exception being passed to
sys.unraisable.