feat: use warnings to report libvips log messages
Previously we directly raised an error when a warning-level log message
was emitted from libvips. This turned out to be too strict, as some
images we process have technically-invalid metadata tags, that cause
VIPS (actually libtiff I think) to emit a warning, and that causes tile
generation to fail.

While this kind of loud failure on unexpected situations is desirable,
we need a way to disable/ignore this specific failure (and other unknown
future problems like it). So to accommodate this without re-inventing
the wheel, we now use the warnings module to emit a warning when such a
libvips warning occurs. By default any such warnings become errors, due
to the default warning filter rules we insert. We ignore the invalid
metadata warnings by default too.

This behaviour can be configured at runtime via the Python warnings
filter controls.